Refactor StepOne/StepOneForm to StepOne
